step 5 - 자동차 초기화 생성 위치를 Racing -> CarFactory로 변경
