+++ src/test/java/racingcar/domain/CarFactoryTest.java
@@ -1,0 +1,18 @@
+package racingcar.domain;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static racingcar.domain.CarFactory.initCars;
+
+public class CarFactoryTest {
+    @Test
+    void 자동차_초기화_성공() {
+        String[] carsArray = {"애플", "삼성", "구글"};
+        List<Car> cars = initCars(carsArray);
+        assertThat(cars).hasSize(3);
+        assertThat(cars).containsExactly(new Car("애플"), new Car("삼성"), new Car("구글"));
+    }
+}
